requirements.txt
Django==1.10
django-gcm==1.2.0
django-tastypie==0.13.3
djangorestframework==3.1.3
mock==1.0.1
python-dateutil==2.5.3
python-mimeparse==1.5.2
pytz==2015.4
requests==2.7.0
six==1.10.0
